Simplify docs build
