+++ docs/core_docs/.gitignore
@@ -31,19 +31,32 @@ yarn-error.log*
 !.yarn/versions
 
 /.quarto/
-
 # AUTO_GENERATED_DOCS
+docs/expression_language/streaming.md
 docs/expression_language/streaming.mdx
+docs/use_cases/tool_use/human_in_the_loop.md
 docs/use_cases/tool_use/human_in_the_loop.mdx
+docs/use_cases/question_answering/streaming.md
 docs/use_cases/question_answering/streaming.mdx
+docs/use_cases/question_answering/sources.md
 docs/use_cases/question_answering/sources.mdx
+docs/use_cases/question_answering/quickstart.md
 docs/use_cases/question_answering/quickstart.mdx
+docs/use_cases/question_answering/local_retrieval_qa.md
 docs/use_cases/question_answering/local_retrieval_qa.mdx
+docs/use_cases/question_answering/conversational_retrieval_agents.md
 docs/use_cases/question_answering/conversational_retrieval_agents.mdx
+docs/use_cases/question_answering/citations.md
 docs/use_cases/question_answering/citations.mdx
+docs/use_cases/question_answering/chat_history.md
 docs/use_cases/question_answering/chat_history.mdx
+docs/modules/model_io/output_parsers/custom.md
 docs/modules/model_io/output_parsers/custom.mdx
-docs/modules/model_io/chat/function_calling_2.mdx
+docs/modules/model_io/chat/function_calling.md
+docs/modules/model_io/chat/function_calling.mdx
+docs/modules/memory/chat_messages/custom.md
 docs/modules/memory/chat_messages/custom.mdx
+docs/modules/data_connection/vectorstores/custom.md
 docs/modules/data_connection/vectorstores/custom.mdx
+docs/modules/model_io/output_parsers/types/openai_tools.md
 docs/modules/model_io/output_parsers/types/openai_tools.mdx
