+++ docs/core_docs/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "docusaurus": "docusaurus",
-    "start": "yarn build:typedoc && rimraf ./docs/api && NODE_OPTIONS=--max-old-space-size=7168 docusaurus start",
+    "start": "yarn build:typedoc && yarn quarto && rimraf ./docs/api && NODE_OPTIONS=--max-old-space-size=7168 docusaurus start",
     "build": "yarn clean && yarn build:typedoc && yarn quarto && rimraf ./build && NODE_OPTIONS=--max-old-space-size=7168 DOCUSAURUS_SSR_CONCURRENCY=4 docusaurus build",
     "build:vercel": "yarn clean && yarn build:typedoc:vercel && yarn quarto:vercel && rimraf ./build && NODE_OPTIONS=--max-old-space-size=7168 DOCUSAURUS_SSR_CONCURRENCY=4 docusaurus build",
     "build:typedoc": "yarn workspace api_refs build",
@@ -22,8 +22,8 @@
     "format": "prettier --write \"**/*.{js,jsx,ts,tsx,md,mdx}\"",
     "format:check": "prettier --check \"**/*.{js,jsx,ts,tsx,md,mdx}\"",
     "clean": "rm -rf .docusaurus/ .turbo/ .build/",
-    "quarto": "quarto render docs/ && node ./scripts/quatro-build.js",
-    "quarto:vercel": "bash ./scripts/vercel_build.sh && node ./scripts/quatro-build.js"
+    "quarto": "quarto render docs/ && node ./scripts/quarto-build.js",
+    "quarto:vercel": "bash ./scripts/vercel_build.sh && node ./scripts/quarto-build.js"
   },
   "dependencies": {
     "@docusaurus/core": "2.4.3",
